main.py

#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
SSV Monetization Tool - Main Entry Point

Инструмент для управления монетизацией контента в SSVproff-Ecosystem.
Автор: Профессор С.В. Сушков
"""

import sys
import yaml
from pathlib import Path


def load_config(config_path: str = "monetization_config.yaml") -> dict:
    """Загрузка конфигурации монетизации."""
    try:
        with open(config_path, 'r', encoding='utf-8') as f:
            config = yaml.safe_load(f)
        return config
    except FileNotFoundError:
        print(f"❌ Файл конфигурации {config_path} не найден.")
        sys.exit(1)
    except yaml.YAMLError as e:
        print(f"❌ Ошибка парсинга YAML: {e}")
        sys.exit(1)


def main():
    """Основная функция инструмента монетизации."""
    print("🚀 SSV Monetization Tool")
    print("=" * 50)
    
    # Загрузка конфигурации
    config = load_config()
    strategy = config.get('monetization', {}).get('strategy', 'hidden')
    
    print(f"📊 Текущая стратегия монетизации: {strategy.upper()}")
    print("=" * 50)
    
    # TODO: Реализация основной логики
    print("\n⚠️  Основная функциональность будет добавлена в следующих версиях.")
    print("📝 Интеграция с ssv-web-dashboard, ssv-video и ssv-book-generator.")
    

if __name__ == "__main__":
    main()
