requirements.txt

pyyaml>=6.0
python-dotenv>=0.19.0
# requests>=2.28.0 # (опционально, для работы с API, например, сокращение ссылок)
# pillow>=9.0.0 # (опционально, если будут изменения в превью)
